Meteorological data

Downloading hourly, daily, and monthly meteorological data from the SYNOP / CLIMATE / PRECIP stations available in the danepubliczne.imgw.pl collection

meteo(interval, rank, year, status = FALSE, coords = FALSE,
  station = NULL, col_names = "short", ...)

Arguments

interval
temporal resolution of the data ("hourly", "daily", "monthly")

rank
rank of the stations ("synop", "climate", "precip")

year
vector of years (e.g., 1966:2000)

status
leave the columns with measurement and observation statuses (default status = FALSE - i.e. the status columns are deleted)

coords
add coordinates of the station (logical value TRUE or FALSE)

station
vector of hydrological stations danepubliczne.imgw.pl can be name of station CAPITAL LETTERS(character) It accepts names (characters in CAPITAL LETTERS) or stations' IDs (numeric)

col_names
three types of column names possible: "short" - default, values with shorten names, "full" - full English description, "polish" - original names in the dataset

...
other parameters that may be passed to the 'shortening' function that shortens column names

Examples

  x <- meteo("monthly", rank = "synop", year = 2018, coords = TRUE)
  head(x)

#>            rank        id        X        Y   station   yy mm tmax_abs
#> 557 SYNOPTYCZNA 353230295 23.16228 53.10726 BIAŁYSTOK 2018  1      7.0
#> 561 SYNOPTYCZNA 353230295 23.16228 53.10726 BIAŁYSTOK 2018  2      3.2
#> 562 SYNOPTYCZNA 353230295 23.16228 53.10726 BIAŁYSTOK 2018  3     12.8
#> 563 SYNOPTYCZNA 353230295 23.16228 53.10726 BIAŁYSTOK 2018  4     27.6
#> 564 SYNOPTYCZNA 353230295 23.16228 53.10726 BIAŁYSTOK 2018  5     29.6
#> 565 SYNOPTYCZNA 353230295 23.16228 53.10726 BIAŁYSTOK 2018  6     30.3
#>     tmax_mean tmin_abs tmin_mean t2m_mean_mon t5cm_min rr_monthly rr_max_daily
#> 557       0.8    -11.5      -3.6         -1.4    -13.3       27.4          7.1
#> 561      -1.8    -18.3      -7.0         -4.5    -18.6       14.9         10.2
#> 562       3.3    -20.2      -5.4         -1.2    -21.8       22.9         10.8
#> 563      18.5     -4.3       5.4         11.7     -6.7       41.0         31.8
#> 564      23.2      4.1       9.5         16.7      0.5       31.2         14.6
#> 565      24.8      2.3      11.3         18.0     -0.6       22.4         10.6
#>     first_day_max_rr last_day_max_rr insolation_monthly snowcover_max
#> 557               29              NA               16.8            18
#> 561                1              NA               47.3             3
#> 562               31              NA              143.9             4
#> 563                1              NA              248.6             4
#> 564               22              NA              363.2             0
#> 565               23              NA                0.0             0
#>     snowcover_days rain_days snow_days r_s_days hail_days fog_days
#> 557             13         5         9        2         0        7
#> 561             18         0        10        4         0        4
#> 562              7         5         7        5         0        7
#> 563              2         5         2        1         0        1
#> 564              0         7         0        0         0        0
#> 565              0         7         0        0         0        0
#>     fogginess_days rime_days glaze_days snowstorm_low_days snowstorm_high_days
#> 557             25         4          3                  2                   1
#> 561             25         1          1                  1                   0
#> 562             27         0          0                  0                   0
#> 563             16         0          0                  0                   0
#> 564              6         0          0                  0                   0
#> 565              4         0          0                  0                   0
#>     hazyness_days ws_10ms_days ws_15ms_days thunder_days dew_days
#> 557             0            0            0            0        1
#> 561             0            0            0            0        0
#> 562             2            0            0            0        0
#> 563             0            0            0            3       14
#> 564             0            0            0            6       22
#> 565             0            0            0            1       16
#>     hoarfrost_days cloud_mean_mon ws_mean_mon vapor_press_mean_mon rh_mean_mon
#> 557             10            6.9         2.8                  5.1        87.1
#> 561              5            6.4         2.1                  3.8        81.5
#> 562             13            5.1         2.3                  4.4        72.7
#> 563              1            3.9         2.6                  9.2        68.0
#> 564              0            2.8         2.2                 11.5        63.7
#> 565              0            4.1         2.0                 12.3        62.2
#>     press_mean_mon slp_mean_mon rr_daytime rr_nightime
#> 557          997.7       1017.0       12.6        14.8
#> 561         1001.2       1020.8        4.0        10.9
#> 562          991.2       1010.3       10.5        12.4
#> 563          996.2       1014.5       21.9        19.1
#> 564         1000.6       1018.7       13.9        17.3
#> 565          996.5       1014.4       16.2         6.2