Hourly meteorological data

Downloading hourly (meteorological) data from the SYNOP / CLIMATE / PRECIP stations available in the danepubliczne.imgw.pl collection

meteo_hourly(rank, year, status = FALSE, coords = FALSE,
  station = NULL, col_names = "short", ...)

Arguments

rank
rank of the stations ("synop", "climate", or "precip")

year
vector of years (e.g., 1966:2000)

status
leave the columns with measurement and observation statuses (default status = FALSE - i.e. the status columns are deleted)

coords
add coordinates of the station (logical value TRUE or FALSE)

station
name or ID of meteorological station(s). It accepts names (characters in CAPITAL LETTERS) or stations' IDs (numeric)

col_names
three types of column names possible: "short" - default, values with shorten names, "full" - full English description, "polish" - original names in the dataset

...
other parameters that may be passed to the 'shortening' function that shortens column names

Examples

  hourly <- meteo_hourly(rank = "climate", year = 1984)
  head(hourly)

#>               rank        id         station   yy mm day hour t2m t_wet
#> 660136 KLIMATYCZNA 249220170 BALIGRÓD-MCHAWA 1984  1   1    0 1.5     0
#> 660137 KLIMATYCZNA 249220170 BALIGRÓD-MCHAWA 1984  1   1    6 1.0     0
#> 660138 KLIMATYCZNA 249220170 BALIGRÓD-MCHAWA 1984  1   1   12 2.3     0
#> 660139 KLIMATYCZNA 249220170 BALIGRÓD-MCHAWA 1984  1   1   18 2.4     0
#> 660140 KLIMATYCZNA 249220170 BALIGRÓD-MCHAWA 1984  1   2    0 1.2     0
#> 660141 KLIMATYCZNA 249220170 BALIGRÓD-MCHAWA 1984  1   2    6 3.4     0
#>        ice_indicator ventil_indicator rh wd_code ws cl_tot visibility_code
#> 660136             W                   0          0      0               0
#> 660137             W                   0      SW  9     10               7
#> 660138             W                   0       S  5      6               7
#> 660139             W                   0     SSW  5      5               7
#> 660140             W                   0          0      0               0
#> 660141             W                   0       S  5      4               7