Monthly meteorological data

Downloading monthly (meteorological) data from the SYNOP / CLIMATE / PRECIP stations available in the danepubliczne.imgw.pl collection

meteo_monthly(rank, year, status = FALSE, coords = FALSE,
  station = NULL, col_names = "short", ...)

Arguments

rank
rank of the stations ("synop", "climate", or "precip")

year
vector of years (e.g., 1966:2000)

status
leave the columns with measurement and observation statuses (default status = FALSE - i.e. the status columns are deleted)

coords
add coordinates of the station (logical value TRUE or FALSE)

station
name or ID of meteorological station(s). It accepts names (characters in CAPITAL LETTERS) or stations' IDs (numeric)

col_names
three types of column names possible: "short" - default, values with shorten names, "full" - full English description, "polish" - original names in the dataset

...
other parameters that may be passed to the 'shortening' function that shortens column names

Examples

  monthly <- meteo_monthly(rank = "climate", year = 1969)
  head(monthly)

#>             rank        id         station   yy mm tmax_abs tmax_mean tmin_abs
#> 1716 KLIMATYCZNA 249220170 BALIGRÓD-MCHAWA 1969  1      9.0      -0.3    -24.2
#> 1720 KLIMATYCZNA 249220170 BALIGRÓD-MCHAWA 1969  2      9.0       1.6    -26.1
#> 1721 KLIMATYCZNA 249220170 BALIGRÓD-MCHAWA 1969  3     10.0       2.2    -17.0
#> 1722 KLIMATYCZNA 249220170 BALIGRÓD-MCHAWA 1969  4     24.5      10.9     -7.4
#> 1723 KLIMATYCZNA 249220170 BALIGRÓD-MCHAWA 1969  5     27.7      20.0     -1.0
#> 1724 KLIMATYCZNA 249220170 BALIGRÓD-MCHAWA 1969  6     25.5      18.8      3.8
#>      tmin_mean t2m_mean_mon t5cm_min rr_monthly rr_max_daily first_day_max_rr
#> 1716     -10.9         -5.5        0       10.8          6.8               24
#> 1720      -5.7         -2.3        0       25.9          6.9                3
#> 1721      -5.6         -1.6        0       25.2          9.2               27
#> 1722      -0.2          5.6        0       56.6         13.0               14
#> 1723       7.1         14.3        0       76.0         36.6               29
#> 1724       9.5         14.4        0      137.6         39.9                8
#>      last_day_max_rr snowcover_max snowcover_days rain_days snow_days
#> 1716              NA            17             24         1         5
#> 1720              NA            13             20         4         5
#> 1721              NA            11             17         2        12
#> 1722              NA            11              9         5         5
#> 1723              NA             0              0        11         0
#> 1724              NA             0              0        19         0
#>      rh_mean_mon ws_mean_mon cloud_mean_mon
#> 1716           0         1.5            4.2
#> 1720           0         1.0            5.6
#> 1721           0         1.1            5.4
#> 1722           0         1.1            3.6
#> 1723           0         1.3            3.6
#> 1724           0         1.2            5.7

  # a descriptive (long) column names:
  monthly2 <- meteo_monthly(rank = "synop", year = 2018, col_names = "full")
  head(monthly2)

#>     Station rank code        ID   Station Year Month
#> 556       SYNOPTYCZNA 353230295 BIAŁYSTOK 2018     1
#> 560       SYNOPTYCZNA 353230295 BIAŁYSTOK 2018     2
#> 561       SYNOPTYCZNA 353230295 BIAŁYSTOK 2018     3
#> 562       SYNOPTYCZNA 353230295 BIAŁYSTOK 2018     4
#> 563       SYNOPTYCZNA 353230295 BIAŁYSTOK 2018     5
#> 564       SYNOPTYCZNA 353230295 BIAŁYSTOK 2018     6
#>     Absolute maximum air temperature [C]
#> 556                                  7.0
#> 560                                  3.2
#> 561                                 12.8
#> 562                                 27.6
#> 563                                 29.6
#> 564                                 30.3
#>     Monthly mean maximum air temperature [C]
#> 556                                      0.8
#> 560                                     -1.8
#> 561                                      3.3
#> 562                                     18.5
#> 563                                     23.2
#> 564                                     24.8
#>     Absolute minimum air temperature [C]
#> 556                                -11.5
#> 560                                -18.3
#> 561                                -20.2
#> 562                                 -4.3
#> 563                                  4.1
#> 564                                  2.3
#>     Monthly mean minimum air temperature [C] Monthly mean air temperature [C]
#> 556                                     -3.6                             -1.4
#> 560                                     -7.0                             -4.5
#> 561                                     -5.4                             -1.2
#> 562                                      5.4                             11.7
#> 563                                      9.5                             16.7
#> 564                                     11.3                             18.0
#>     Minimum daily near surface air temperature [C]
#> 556                                          -13.3
#> 560                                          -18.6
#> 561                                          -21.8
#> 562                                           -6.7
#> 563                                            0.5
#> 564                                           -0.6
#>     Total monthly precipitation [mm] Maximum daily preciptation [mm]
#> 556                             27.4                             7.1
#> 560                             14.9                            10.2
#> 561                             22.9                            10.8
#> 562                             41.0                            31.8
#> 563                             31.2                            14.6
#> 564                             22.4                            10.6
#>     First day with maximum precipitation Last day with maximum precipitation
#> 556                                   29                                  NA
#> 560                                    1                                  NA
#> 561                                   31                                  NA
#> 562                                    1                                  NA
#> 563                                   22                                  NA
#> 564                                   23                                  NA
#>     Total monthly insolation [hours] Maximum daily snow cover [cm]
#> 556                             16.8                            18
#> 560                             47.3                             3
#> 561                            143.9                             4
#> 562                            248.6                             4
#> 563                            363.2                             0
#> 564                              0.0                             0
#>     Days with snow cover Days with rain Days with snow Days with sleet
#> 556                   13              5              9               2
#> 560                   18              0             10               4
#> 561                    7              5              7               5
#> 562                    2              5              2               1
#> 563                    0              7              0               0
#> 564                    0              7              0               0
#>     Days with hail Days with fog Days with fogginess Days with rime
#> 556              0             7                  25              4
#> 560              0             4                  25              1
#> 561              0             7                  27              0
#> 562              0             1                  16              0
#> 563              0             0                   6              0
#> 564              0             0                   4              0
#>     Days with glaze Days with low snowstorm Days with high snowstorm
#> 556               3                       2                        1
#> 560               1                       1                        0
#> 561               0                       0                        0
#> 562               0                       0                        0
#> 563               0                       0                        0
#> 564               0                       0                        0
#>     Days with hazyness Days with wind speed above 10 ms
#> 556                  0                                0
#> 560                  0                                0
#> 561                  2                                0
#> 562                  0                                0
#> 563                  0                                0
#> 564                  0                                0
#>     Days with wind speed above 15 ms Days with thunderstorm Days with dew
#> 556                                0                      0             1
#> 560                                0                      0             0
#> 561                                0                      0             0
#> 562                                0                      3            14
#> 563                                0                      6            22
#> 564                                0                      1            16
#>     Days with hoarfrost Monthly mean total cloudiness [hPa]
#> 556                  10                                 6.9
#> 560                   5                                 6.4
#> 561                  13                                 5.1
#> 562                   1                                 3.9
#> 563                   0                                 2.8
#> 564                   0                                 4.1
#>     Monthly mean wind speed [m/s] Monthly mean vapour pressure [hPa]
#> 556                           2.8                                5.1
#> 560                           2.1                                3.8
#> 561                           2.3                                4.4
#> 562                           2.6                                9.2
#> 563                           2.2                               11.5
#> 564                           2.0                               12.3
#>     Monthly mean relative humidity [%]
#> 556                               87.1
#> 560                               81.5
#> 561                               72.7
#> 562                               68.0
#> 563                               63.7
#> 564                               62.2
#>     Monthly mean pressure at station level [hPa]
#> 556                                        997.7
#> 560                                       1001.2
#> 561                                        991.2
#> 562                                        996.2
#> 563                                       1000.6
#> 564                                        996.5
#>     Monthly mean sea level pressure [hPa] Total precipitation during day [mm]
#> 556                                1017.0                                12.6
#> 560                                1020.8                                 4.0
#> 561                                1010.3                                10.5
#> 562                                1014.5                                21.9
#> 563                                1018.7                                13.9
#> 564                                1014.4                                16.2
#>     Total precipitation during night [mm]
#> 556                                  14.8
#> 560                                  10.9
#> 561                                  12.4
#> 562                                  19.1
#> 563                                  17.3
#> 564                                   6.2